# message_service/docs/swagger.yaml
basePath: /
definitions:
...
  controllers.AddUserToChatRequest:
    properties:
      user_id:
        type: integer
    required:
    - user_id
    type: object
...
  controllers.CloseChatResponse:
    properties:
      message:
        type: string
    type: object
  controllers.CreateChatRequest:
    properties:
      assigned_to:
        type: integer
      name:
        type: string
      users:
        description: Массив ID пользователей
        items:
          type: integer
        type: array
    required:
    - name
    - users
    type: object
  controllers.CreateChatResponse:
    properties:
      chat_id:
        type: integer
      created_at:
        type: string
      message:
        type: string
      name:
        type: string
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
...
  controllers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  controllers.GetChatHistoryResponse:
    properties:
      messages:
        items:
          $ref: '#/definitions/models.Message'
        type: array
    type: object
  controllers.GetChatResponse:
    properties:
      chat:
        $ref: '#/definitions/models.Chat'
    type: object
...
  controllers.RemoveUserFromChatRequest:
    properties:
      user_id:
        type: integer
    required:
    - user_id
    type: object
...
  controllers.SuccessResponse:
    properties:
      message:
        type: string
    type: object
  controllers.TransferChatRequest:
    properties:
      new_user_id:
        type: integer
    required:
    - new_user_id
    type: object
...
  models.Chat:
    properties:
      assigned_to:
        type: integer
      created_at:
        type: string
      entry_point:
        type: string
      id:
        type: integer
      messages:
        items:
          $ref: '#/definitions/models.Message'
        type: array
      name:
        type: string
      status:
        type: string
      updated_at:
        type: string
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
  models.Message:
    properties:
      attached_id:
        type: integer
      chat_id:
        type: integer
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      is_checked:
        type: boolean
      is_read:
        type: boolean
      user_id:
        type: integer
    type: object
  models.Role:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      password:
        type: string
      role:
        allOf:
        - $ref: '#/definitions/models.Role'
        description: Связь с моделью Role
      role_id:
        type: integer
      updated_at:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
...
  /chat:
    post:
      consumes:
      - application/json
      description: Этот эндпоинт создает новый чат и добавляет пользователей в него.
      parameters:
      - description: Данные для создания чата
        in: body
        name: chat
        required: true
        schema:
          $ref: '#/definitions/controllers.CreateChatRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.CreateChatResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Создание нового чата
      tags:
      - Чаты
  /chat/{id}:
    get:
      description: Этот эндпоинт возвращает информацию о чате по его ID.
      parameters:
      - description: ID чата
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.GetChatResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Получение информации о чате
      tags:
      - Чаты
  /chat/{id}/add_user:
    post:
      consumes:
      - application/json
      description: Этот эндпоинт добавляет пользователя в чат.
      parameters:
      - description: ID чата
        in: path
        name: id
        required: true
        type: integer
      - description: Информация о пользователе
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/controllers.AddUserToChatRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Добавление пользователя в чат
      tags:
      - Чаты
  /chat/{id}/close:
    post:
      description: Этот эндпоинт закрывает чат.
      parameters:
      - description: ID чата
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.CloseChatResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Закрытие чата
      tags:
      - Чаты
  /chat/{id}/history:
    get:
      description: Этот эндпоинт возвращает историю сообщений в чате по его ID.
      parameters:
      - description: ID чата
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.GetChatHistoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Получение истории сообщений в чате
      tags:
      - Чаты
  /chat/{id}/remove_user:
    post:
      consumes:
      - application/json
      description: Этот эндпоинт удаляет пользователя из чата.
      parameters:
      - description: ID чата
        in: path
        name: id
        required: true
        type: integer
      - description: Информация о пользователе
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/controllers.RemoveUserFromChatRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Удаление пользователя из чата
      tags:
      - Чаты
  /chat/{id}/transfer:
    post:
      consumes:
      - application/json
      description: Этот эндпоинт передает чат другому пользователю.
      parameters:
      - description: ID чата
        in: path
        name: id
        required: true
        type: integer
      - description: Данные для передачи чата
        in: body
        name: transfer
        required: true
        schema:
          $ref: '#/definitions/controllers.TransferChatRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Передача чата другому пользователю
      tags:
      - Чаты
...
swagger: "2.0"